+++ packages/ci-tools/vitest.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from '@loop-infra/vitest-config';
+
+export default defineConfig({
+  dirname: __dirname,
+  preset: 'node',
+});