+++ packages/ci-tools/CHANGELOG.md
@@ -1,4 +1,11 @@
 # 🕗 Change Log - @cozeloop/ci-tools
 
+## 0.0.2
+* [feat] add `sync-pr` to synchronize GitHub PR via lark in `lark` command
+
 ## 0.0.1
 🌱 Initial version
+
+`cozeloop-ci`: 🔧 Cozeloop CI tools.
+- `lark`: Lark tools
+  - `sync-issue`: Synchronize GitHub issue via lark